+++ pages/kodex/index.tsx
@@ -6,6 +6,7 @@ import {fetchMenuPosts} from 'utils/fetchMenuPosts'
 import {PostOrPage} from '@tryghost/content-api'
 import {useMenuData} from 'context/SessionContext'
 import {NextSeo} from "next-seo";
+import google from 'utils/google'
 
 // Fetch posts
 export const getServerSideProps: GetServerSideProps = async (context) => {
@@ -26,9 +27,11 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 }
 
 const Kodex = ({
-  menuPosts
+  menuPosts,
+  isCoockiesEnabled
 }: {
 menuPosts?: PostOrPage[]
+isCoockiesEnabled: boolean
 }) => {
   useMenuData({menuPosts})
 
@@ -37,18 +40,7 @@ menuPosts?: PostOrPage[]
       <Head>
         <title>Bankless | Kodex</title>
         <link rel="icon" type="image/png" href="/favicon.png" />
-
-        <script async src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_KEY}`}></script>
-        <script
-          async
-          dangerouslySetInnerHTML={{
-            __html: `window.dataLayer = window.dataLayer || [];
-              function gtag(){dataLayer.push(arguments);}
-              gtag('js', new Date());
-            
-              gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_KEY}');`
-          }}
-        />
+        {google(isCoockiesEnabled)}
       </Head>
       <NextSeo
         title="Bankless | Kodex"
