+++ pages/studium/[id].tsx
@@ -9,6 +9,7 @@ import { fetchMenuPosts } from 'utils/fetchMenuPosts'
 import { useMenuData } from 'context/SessionContext'
 import MetaTags from "../../components/MetaTags/MetaTags";
 import {NextSeo} from "next-seo";
+import google from 'utils/google'
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const { query } = context
@@ -70,11 +71,13 @@ export default function ZacatecniciArticleDetail({
   moreStories,
   hashovky,
   menuPosts,
+  isCoockiesEnabled
 }: {
   articlePost?: PostOrPage
   moreStories?: PostOrPage[]
   hashovky?: PostOrPage[]
   menuPosts?: PostOrPage[]
+  isCoockiesEnabled: boolean
 }) {
   const [hashovkyState, setHashovkyState] = useState<PostOrPage[]>([])
   const articleData = articlePost
@@ -95,17 +98,7 @@ export default function ZacatecniciArticleDetail({
       <link rel="icon" type="image/png" href="/favicon.png" />
 
       <base target="_blank"/>
-      <script async src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_KEY}`}></script>
-      <script
-        async
-        dangerouslySetInnerHTML={{
-          __html: `window.dataLayer = window.dataLayer || [];
-              function gtag(){dataLayer.push(arguments);}
-              gtag('js', new Date());
-            
-              gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_KEY}');`
-        }}
-      />
+      {google(isCoockiesEnabled)}
     </Head>
     <NextSeo
       title={articlePost?.meta_title || ''}
