+++ pages/studium/polkadot/index.tsx
@@ -13,6 +13,7 @@ import { getPosts } from '../../api/posts'
 import styles from '../../../styles/Home.module.scss'
 import { useMenuData } from 'context/SessionContext'
 import {NextSeo} from "next-seo";
+import google from 'utils/google'
 
 export const POSTS_ON_PAGE_LIMIT = 15
 
@@ -56,10 +57,12 @@ const ZacatecniciBitcoin = ({
   posts,
   hashovky,
   menuPosts,
+  isCoockiesEnabled
 }: {
   posts?: PostOrPage[]
   hashovky?: PostOrPage[]
   menuPosts?: PostOrPage[]
+  isCoockiesEnabled: boolean
 }) => {
   const [postsState, setPostsState] = useState<PostOrPage[]>([])
   const [hashovkyState, setHashovkyState] = useState<PostOrPage[]>([])
@@ -86,20 +89,7 @@ const ZacatecniciBitcoin = ({
         <link rel="icon" type="image/png" href="/favicon.png" />
 
         <base target="_blank" />
-        <script
-          async
-          src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_KEY}`}
-        ></script>
-        <script
-          async
-          dangerouslySetInnerHTML={{
-            __html: `window.dataLayer = window.dataLayer || [];
-              function gtag(){dataLayer.push(arguments);}
-              gtag('js', new Date());
-            
-              gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_KEY}');`,
-          }}
-        />
+        {google(isCoockiesEnabled)}
       </Head>
       <NextSeo
         title="Bankless | Polkadot pro začátečníky"
