+++ utils/google.tsx
@@ -1,0 +1,17 @@
+export default  function google (isCoockiesEnabled: boolean) {
+  if(isCoockiesEnabled){
+    return (
+      <div>
+        <script async src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_KEY}`}></script>
+        <script
+        async
+        dangerouslySetInnerHTML={{
+          __html: `window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+          
+            gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_KEY}');`
+        }}/>
+      </div>)}
+return (null)
+}
